+++ src/main/java/com/dalo/patterns/behavioralPatterns/chainOfResponsibility/dao/UserDao.java
@@ -1,0 +1,13 @@
+package com.dalo.patterns.behavioralPatterns.chainOfResponsibility.dao;
+
+import com.dalo.patterns.behavioralPatterns.chainOfResponsibility.model.User;
+
+public class UserDao {
+    public User getUserByEmail(String email) {
+        return null;
+    }
+
+    public User createUser(User user) {
+        return user;
+    }
+}
