+++ src/main/java/com/dalo/patterns/behavioralPatterns/chainOfResponsibility/middleware/Middleware.java
@@ -1,0 +1,20 @@
+package com.dalo.patterns.behavioralPatterns.chainOfResponsibility.middleware;
+
+public abstract class Middleware {
+    private Middleware next;
+
+    public Middleware linkWith(Middleware next) {
+        this.next = next;
+        return next;
+    }
+
+    protected boolean checkNext(String email, String password) {
+        if (next == null) {
+            return true;
+        }
+        return next.check(email, password);
+    }
+
+    public abstract boolean check(String email, String password);
+
+}
