+++ src/main/java/com/dalo/patterns/behavioralPatterns/chainOfResponsibility/model/User.java
@@ -1,0 +1,16 @@
+package com.dalo.patterns.behavioralPatterns.chainOfResponsibility.model;
+
+public class User {
+    private Long id;
+    private String email;
+    private String password;
+
+    public User() {
+
+    }
+
+    public User(String email, String password) {
+        this.email = email;
+        this.password = email;
+    }
+}
