+++ src/main/java/com/dalo/patterns/behavioralPatterns/command/command/Command.java
@@ -1,0 +1,5 @@
+package com.dalo.patterns.behavioralPatterns.command.command;
+
+public abstract class Command {
+    public abstract boolean execute();
+}
