+++ src/main/java/com/dalo/patterns/behavioralPatterns/command/command/CommandHistory.java
@@ -1,0 +1,19 @@
+package com.dalo.patterns.behavioralPatterns.command.command;
+
+import java.util.Stack;
+
+public class CommandHistory {
+    private Stack<Command> commandHistory = new Stack<>();
+
+    public void push(Command command) {
+        commandHistory.push(command);
+    }
+
+    public Command pop() {
+        return commandHistory.pop();
+    }
+
+    public boolean isEmpty() {
+        return commandHistory.isEmpty();
+    }
+}
