+++ src/main/java/com/dalo/patterns/behavioralPatterns/command/command/commands/createCommand.java
@@ -1,0 +1,10 @@
+package com.dalo.patterns.behavioralPatterns.command.command.commands;
+
+import com.dalo.patterns.behavioralPatterns.command.command.Command;
+
+public class createCommand extends Command {
+    @Override
+    public boolean execute() {
+        return false;
+    }
+}
