+++ src/main/java/com/dalo/patterns/behavioralPatterns/iterator/iterator/EntityCollection.java
@@ -1,0 +1,5 @@
+package com.dalo.patterns.behavioralPatterns.iterator.iterator;
+
+public interface EntityCollection {
+    public EntityIterator getIterator();
+}
