+++ src/main/java/com/dalo/patterns/behavioralPatterns/iterator/iterator/EntityIterator.java
@@ -1,0 +1,9 @@
+package com.dalo.patterns.behavioralPatterns.iterator.iterator;
+
+import com.dalo.patterns.behavioralPatterns.iterator.model.Entity;
+
+public interface EntityIterator {
+    boolean hasNext();
+    Entity getNext();
+    void reset();
+}
