+++ src/main/java/com/dalo/patterns/behavioralPatterns/iterator/model/Entity.java
@@ -1,0 +1,6 @@
+package com.dalo.patterns.behavioralPatterns.iterator.model;
+
+public class Entity {
+    private Long id;
+    private String criteria;
+}
