+++ src/main/java/com/dalo/patterns/behavioralPatterns/observer/event/EventListener.java
@@ -1,0 +1,5 @@
+package com.dalo.patterns.behavioralPatterns.observer.event;
+
+public interface EventListener {
+    void update(String eventType);
+}
