+++ src/main/java/com/dalo/patterns/behavioralPatterns/observer/event/impl/NotificationListener.java
@@ -1,0 +1,10 @@
+package com.dalo.patterns.behavioralPatterns.observer.event.impl;
+
+import com.dalo.patterns.behavioralPatterns.observer.event.EventListener;
+
+public class NotificationListener implements EventListener {
+    @Override
+    public void update(String something) {
+        System.out.println(something);
+    }
+}
