+++ src/main/java/com/dalo/patterns/behavioralPatterns/strategy/strategy/ExceptionRule.java
@@ -1,0 +1,5 @@
+package com.dalo.patterns.behavioralPatterns.strategy.strategy;
+
+public interface ExceptionRule {
+    void resolve();
+}
