+++ src/main/java/com/dalo/patterns/behavioralPatterns/strategy/strategy/impl/HashTableException.java
@@ -1,0 +1,10 @@
+package com.dalo.patterns.behavioralPatterns.strategy.strategy.impl;
+
+import com.dalo.patterns.behavioralPatterns.strategy.strategy.ExceptionRule;
+
+public class HashTableException implements ExceptionRule {
+    @Override
+    public void resolve() {
+        System.out.println("HashTable");
+    }
+}
