+++ src/main/java/com/dalo/patterns/behavioralPatterns/strategy/strategy/impl/PostgresException.java
@@ -1,0 +1,10 @@
+package com.dalo.patterns.behavioralPatterns.strategy.strategy.impl;
+
+import com.dalo.patterns.behavioralPatterns.strategy.strategy.ExceptionRule;
+
+public class PostgresException implements ExceptionRule {
+    @Override
+    public void resolve() {
+        System.out.println("Postgres");
+    }
+}
