+++ src/main/java/com/dalo/patterns/behavioralPatterns/templateMethod/template/FirstNetwork.java
@@ -1,0 +1,9 @@
+package com.dalo.patterns.behavioralPatterns.templateMethod.template;
+
+public class FirstNetwork extends Network {
+    @Override
+    boolean logIn(String userName, String password) {
+        System.out.println("first");
+        return true;
+    }
+}
