+++ src/main/java/com/dalo/patterns/behavioralPatterns/templateMethod/template/SecondNetwork.java
@@ -1,0 +1,10 @@
+package com.dalo.patterns.behavioralPatterns.templateMethod.template;
+
+public class SecondNetwork extends Network {
+
+    @Override
+    boolean logIn(String userName, String password) {
+        System.out.println("second");
+        return true;
+    }
+}
