--- src/main/java/com/dalo/patterns/behavioral_patterns/chain_of_responsibility/dao/UserDao.java
@@ -1,13 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.chain_of_responsibility.dao;
-
-import com.dalo.patterns.behavioral_patterns.chain_of_responsibility.model.User;
-
-public class UserDao {
-    public User getUserByEmail(String email) {
-        return null;
-    }
-
-    public User createUser(User user) {
-        return user;
-    }
-}
