--- src/main/java/com/dalo/patterns/behavioral_patterns/chain_of_responsibility/middleware/Middleware.java
@@ -1,20 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.chain_of_responsibility.middleware;
-
-public abstract class Middleware {
-    private Middleware next;
-
-    public Middleware linkWith(Middleware next) {
-        this.next = next;
-        return next;
-    }
-
-    protected boolean checkNext(String email, String password) {
-        if (next == null) {
-            return true;
-        }
-        return next.check(email, password);
-    }
-
-    public abstract boolean check(String email, String password);
-
-}
