--- src/main/java/com/dalo/patterns/behavioral_patterns/command/command/Command.java
@@ -1,5 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.command.command;
-
-public abstract class Command {
-    public abstract boolean execute();
-}
