--- src/main/java/com/dalo/patterns/behavioral_patterns/command/command/CommandHistory.java
@@ -1,19 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.command.command;
-
-import java.util.Stack;
-
-public class CommandHistory {
-    private Stack<Command> commandHistory = new Stack<>();
-
-    public void push(Command command) {
-        commandHistory.push(command);
-    }
-
-    public Command pop() {
-        return commandHistory.pop();
-    }
-
-    public boolean isEmpty() {
-        return commandHistory.isEmpty();
-    }
-}
