--- src/main/java/com/dalo/patterns/behavioral_patterns/command/command/commands/deleteCommand.java
@@ -1,10 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.command.command.commands;
-
-import com.dalo.patterns.behavioral_patterns.command.command.Command;
-
-public class deleteCommand extends Command {
-    @Override
-    public boolean execute() {
-        return false;
-    }
-}
