--- src/main/java/com/dalo/patterns/behavioral_patterns/iterator/iterator/EntityCollection.java
@@ -1,5 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.iterator.iterator;
-
-public interface EntityCollection {
-    public EntityIterator getIterator();
-}
