--- src/main/java/com/dalo/patterns/behavioral_patterns/iterator/iterator/EntityIterator.java
@@ -1,9 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.iterator.iterator;
-
-import com.dalo.patterns.behavioral_patterns.iterator.model.Entity;
-
-public interface EntityIterator {
-    boolean hasNext();
-    Entity getNext();
-    void reset();
-}
