--- src/main/java/com/dalo/patterns/behavioral_patterns/iterator/model/Entity.java
@@ -1,6 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.iterator.model;
-
-public class Entity {
-    private Long id;
-    private String criteria;
-}
