--- src/main/java/com/dalo/patterns/behavioral_patterns/observer/event/EventListener.java
@@ -1,5 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.observer.event;
-
-public interface EventListener {
-    void update(String eventType);
-}
