--- src/main/java/com/dalo/patterns/behavioral_patterns/observer/service/Service.java
@@ -1,23 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.observer.service;
-
-import com.dalo.patterns.behavioral_patterns.observer.event.EventListener;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Service {
-    private List<EventListener> observers = new ArrayList<>();
-    public void addObserver(EventListener eventListener) {
-        observers.add(eventListener);
-    }
-    public void removeObserver(EventListener eventListener) {
-        observers.remove(eventListener);
-    }
-    public void updateObserver(String something) {
-        for (EventListener eventListener : this.observers) {
-            eventListener.update(something);
-        }
-    }
-
-
-}
