--- src/main/java/com/dalo/patterns/behavioral_patterns/strategy/strategy/ExceptionRule.java
@@ -1,5 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.strategy.strategy;
-
-public interface ExceptionRule {
-    void resolve();
-}
