--- src/main/java/com/dalo/patterns/behavioral_patterns/strategy/strategy/impl/PostgresException.java
@@ -1,10 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.strategy.strategy.impl;
-
-import com.dalo.patterns.behavioral_patterns.strategy.strategy.ExceptionRule;
-
-public class PostgresException implements ExceptionRule {
-    @Override
-    public void resolve() {
-        System.out.println("Postgres");
-    }
-}
