--- src/main/java/com/dalo/patterns/behavioral_patterns/template_method/template/FirstNetwork.java
@@ -1,9 +1,0 @@
-package com.dalo.patterns.behavioral_patterns.template_method.template;
-
-public class FirstNetwork extends Network {
-    @Override
-    boolean logIn(String userName, String password) {
-        System.out.println("first");
-        return true;
-    }
-}
